backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.3.0
openai==1.12.0
